Hook up isLoading to profile component.
